corrections
gcpsa and gcpsb corrected,
various small changes,
everything works fine now

+++ gcpscode.c
@@ -28,19 +28,31 @@ partial_alloc simple_GCPS_alloc(input_sch_ch_prob* input) {
 partial_alloc GCPS_allocation(input_sch_ch_prob* input, run_data* data) {
   partial_alloc feasible_guide;
   pivot_list probe_list;
-  input_sch_ch_prob stu_no_pr_scp;
   process_scp input_pr_scp;
   pure_alloc pure_da;
+
+  if (!safe_schools_are_safe(input)) {
+    fprintf(stderr, "Unsafe safe schools in input!\n");
+    exit(0);
+  }
+  else {
+    fprintf(stderr, "input is OK.\n");
+  }
+
+  fprintf(stderr, "Going into deferred acceptannce.\n");
   
-  stu_no_pr_scp = stu_no_priority_scp(input);
-  input_pr_scp = process_scp_from_input(input);
-  pure_da = deferred_acceptance(&stu_no_pr_scp); 
+  pure_da = deferred_acceptance(input); 
+
+  fprintf(stderr, "Leaving deferred acceptannce.\n");
+  
   feasible_guide = partial_allocation_from_pure(&pure_da);
   
-  destroy_input_sch_ch_prob(stu_no_pr_scp);
+  /*  destroy_input_sch_ch_prob(stu_no_pr_scp); */
   destroy_pure_alloc(pure_da);  
   destroy_input_sch_ch_prob(*input);
   
+  
+  input_pr_scp = process_scp_from_input(input);
   probe_list = void_pivot_list();
 
   return GCPS_allocation_with_guide(&input_pr_scp, &feasible_guide, &probe_list, data);
